логика отделена от представления
