+++ js/game.js
@@ -1,37 +1,19 @@
-import {changePageTemplate} from './utils.js';
-import {getData, getState, setState} from './data';
-import {getGameModule} from './gameView';
-import stats from './stats';
+import {getState, setState} from './data';
 
-const stateModule = {question1: ``, question2: ``};
+const stateModule = { question1: ``, question2: `` };
 
-function getModule() {
-  const levels = getData().levels;
-  const curLevel = getState().curLevel;
-  if (curLevel < levels.length) {
-    return getGameModule(levels[curLevel], getState());
-  } else {
-    return stats;
-  }
-}
-
-export function nextModule() {
+export function nextLevel() {
   setState(`curLevel`, getState().curLevel + 1);
-  return getModule();
 }
 
 export function checkComplete(e) {
   stateModule[e.currentTarget.name] = e.currentTarget.value;
-  let complete = true;
   let key;
+  let complete = true;
   for (key in stateModule) {
     if (!stateModule[key]) {
       complete = false;
     }
   }
-  if (complete) {
-    changePageTemplate(nextModule());
-  }
+  return complete;
 }
-
-export default getModule();
